Add missing scanner.c to README and Package.swift
Prior to this change, a new scanner.c source file had been added to the
repository. The Makefile had been updated, but the command in the README
was missing the new file.

This change adds the new file to the command in the README, and also
updates the Package.swift file to declare it as a source.